pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llamonitor-async"
version = "0.2.0"
description = "Lightweight async monitoring for LLM applications - capacity-based tracking with pluggable storage"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "LLMOps Monitoring Contributors"}
]
keywords = ["llm", "monitoring", "observability", "async", "telemetry"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "pydantic>=2.0.0",
    "aiofiles>=23.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
parquet = [
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",
]
postgres = [
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.0",
]
mysql = [
    "aiomysql>=0.2.0",
    "sqlalchemy>=2.0.0",
]
clickhouse = [
    "clickhouse-driver>=0.2.6",
]
tokens = [
    "tiktoken>=0.5.0",
]
prometheus = [
    "prometheus-client>=0.19.0",
]
api = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
]
graphql = [
    "strawberry-graphql[fastapi]>=0.216.0",
]
datadog = [
    "datadog-api-client>=2.0.0",
    "datadog>=0.47.0",
]
all = [
    "llamonitor-async[parquet,postgres,mysql,clickhouse,tokens,prometheus,api,graphql,datadog]",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]

[project.urls]
Homepage = "https://github.com/guybass/LLMOps_monitoring_async-"
Documentation = "https://github.com/guybass/LLMOps_monitoring_async-/blob/main/README.md"
Repository = "https://github.com/guybass/LLMOps_monitoring_async-"
Issues = "https://github.com/guybass/LLMOps_monitoring_async-/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["llmops_monitoring*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["llmops_monitoring/tests"]
